Fix classes to standalone class files. Removed dead test classes/stubs.
