+++ tests/test_app/TestApp/Model/Table/SecondaryPostsTable.php
@@ -1,0 +1,27 @@
+<?php
+declare(strict_types=1);
+/**
+ * CakePHP(tm) : Rapid Development Framework (https://cakephp.org)
+ * Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ *
+ * Licensed under The MIT License
+ * Redistributions of files must retain the above copyright notice
+ *
+ * @copyright     Copyright (c) Cake Software Foundation, Inc. (https://cakefoundation.org)
+ * @since         3.0.0
+ * @license       https://opensource.org/licenses/mit-license.php MIT License
+ */
+namespace TestApp\Model\Table;
+
+use Cake\ORM\Table;
+
+class SecondaryPostsTable extends Table
+{
+    /**
+     * @return string
+     */
+    public static function defaultConnectionName(): string
+    {
+        return 'secondary';
+    }
+}
